保持播放状态

+++ lib/cfgloader.go
@@ -15,8 +15,8 @@ type IptvConfig struct {
 }
 
 type AppConfig struct {
-	Tasktime int8
-	Tasknum int8
+	Tasktime int
+	Tasknum int
 }
 
 type CASUser struct {
